Change type of price to BigDecimal

+++ src/main/java/com/pragmaticcoders/checkout/calulator/PriceCalculator.java
@@ -2,6 +2,8 @@ package com.pragmaticcoders.checkout.calulator;
 
 import com.pragmaticcoders.checkout.domain.Item;
 
+import java.math.BigDecimal;
+
 public interface PriceCalculator {
-    Integer calcCostOfItem(Item item, Integer quantity) throws InvalidPriceException;
+    BigDecimal calcCostOfItem(Item item, Integer quantity) throws InvalidPriceException;
 }
