+++ src/test/java/com/pragmaticcoders/checkout/command/promotion/AddPromotionCommandTest.java
@@ -3,6 +3,7 @@ package com.pragmaticcoders.checkout.command.promotion;
 import com.pragmaticcoders.checkout.dto.PromotionDto;
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.UUID;
 
@@ -13,7 +14,7 @@ public class AddPromotionCommandTest {
     @Test
     public void testCreateAndGet() {
         UUID uuid = UUID.randomUUID();
-        PromotionDto dto = new PromotionDto(new HashSet<>(), 21);
+        PromotionDto dto = new PromotionDto(new HashSet<>(), BigDecimal.valueOf(21));
 
         AddPromotionCommand command = new AddPromotionCommand(uuid, dto);
         assertEquals(uuid, command.getUuid());
