+++ src/test/java/com/pragmaticcoders/checkout/dto/ItemDtoTest.java
@@ -2,6 +2,7 @@ package com.pragmaticcoders.checkout.dto;
 
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 import static org.junit.Assert.assertEquals;
@@ -13,12 +14,12 @@ public class ItemDtoTest {
         ItemDto dto = new ItemDto(
             "name",
             new ArrayList<ItemDto.Price>() {{
-                add(new ItemDto.Price(1, 3));
+                add(new ItemDto.Price(1, BigDecimal.valueOf(3)));
             }}
         );
 
         assertEquals("name", dto.getName());
         assertEquals(Integer.valueOf(1), dto.getPrices().get(0).getQuantity());
-        assertEquals(Integer.valueOf(3), dto.getPrices().get(0).getPrice());
+        assertEquals(BigDecimal.valueOf(3), dto.getPrices().get(0).getPrice());
     }
 }
